GTK3: Support custom panel themes and transparency
GTK3: Support custom panel themes and transparency.
This adds support for custom panel themes using the .mate-panel-menu-bar selector to control child widgets. The menus are controllled by .mate-panel-menu-bar .menu and most of the slider theme is that of the frame it is drawn in. The .mate-panel-menu-bar .frame selector will theme this frame, there is also an inner frame in the layout which will follow .mate-panel-menu-bar .frame .frame and otherwise display a second time anything specified in .mate-panel-menu-bar .frame . In most cases that can simply be set transparent.

Themes can use alpha values for both menus and slider backgrounds.