GTK3: fix volume slider position in gtk3.19
In gtk3.19, the first time you click on the volume control, the slider appears at the top left of the screen. Second and subsequent clicks OK. This patch fixes that and restores the volume slider to always appearing over or under the volume control icon. Tested with gtk3.14 and with gtk3.19.5  GIT master Dec 27 2015, and with both top and bottom panel/tray locations.