applet: fix minor rounding mistakes
The chosen volume step increment isn't necessarily an integer. The
volume however is stored as integer. Whenever the volume is synced back
from the mixer decimal digits may be cut off behind the dot. This
commit tries to counteract the effect by rounding to hundreths of the
maximum volume when syncing back from the mixer.

As the volume is shown in percent the new behavior should be more
reasonable to the user. There are no more irregular jumps and there
is no additional 99% step.