Ignore PulseAudio "monitor" inputs when displaying mic icon
Teradici PCoIP on Linux presently has audio output but no audio input.
PulseAudio creates a "monitor" input device for each output device. These
combine so that a microphone icon is displayed in the tray even though no
actual input device is present. It's misleading, so we created this patch to
avoid it.