mate-volume-control-applet: Use symbolic icons when available.
Use the following icon names when available:

    "audio-status-volume-muted-symbolic",
    "audio-status-volume-low-symbolic",
    "audio-status-volume-medium-symbolic",
    "audio-status-volume-high-symbolic",
    "audio-status-microphone-muted-symbolic",
    "audio-status-microphone-low-symbolic",
    "audio-status-microphone-medium-symbolic",
    "audio-status-microphone-high-symbolic".

If these icons are not present in the icon theme, fallback to
fullcolor icons with legacy names (the same icons that were used
before):

    "audio-volume-muted",
    "audio-volume-low",
    "audio-volume-medium",
    "audio-volume-high",
    "audio-input-microphone-muted",
    "audio-input-microphone-low",
    "audio-input-microphone-medium",
    "audio-input-microphone-high".

Symbolic icons are the only way for themes to properly support
both light and dark panels with grey icons. If the icon name ends
with -symbolic, GTK renders the foreground color of the icon based
on the color of the panel behind it.

It's also important to use different icon names when it comes to
status icons, and not mix them with names which are also used for
notifications, menu items, and other aspects of the desktop.

There are two reasons for this:

- statusicons need to be rendered in specific sizes (typically, the presence
    of a scalable icon would create sizing issues in the systray, but could
    be required for notifications).

- full color icons might need to be rendered differently when notifications,
    menus and panels have a different shade within the same GTK theme (for
    instance with a light menu and a dark panel).

This also makes it more flexible for themes. They have the ability to
create symlinks if they want to, or to provide different icons for
panel, notifications and menus.

Ideally we should change the name of the full color icons as well,
but that would break support for existing themes, so I didn't do this
in this commit.